+++ bitfusion_device_plugin/webhook/deployment/bitfusion-injector.yaml
@@ -19,6 +19,9 @@ spec:
         - name: bwki
           image: phaedobf/sidecar-injector:v0.1
           imagePullPolicy: IfNotPresent
+          env:
+            - name: TOTAL_GPU_MEMORY
+              value: 16000
           args:
           - -sidecarCfgFile=/etc/webhook/config/sidecarconfig.yaml
           - -bitfusionClientConfig=/etc/webhook/bitfusion-client-config/bitfusion-client-config.yaml
